Fix no open position dropdown
