+++ components/Portfolio/Trading/Overview/index.tsx
@@ -6,7 +6,7 @@ import Tracer, { defaults } from '@libs/Tracer';
 import styled from 'styled-components';
 import ConnectOverlay from '@components/Overlay/ConnectOverlay';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
-import PositionOverlay from '@components/Overlay/PositionOverlay';
+import PositionMarketOverlay from '@components/Overlay/PositionMarketOverlay';
 import {
     VScrollContainer,
     PortfolioDropdown,
@@ -17,13 +17,15 @@ import {
     HPanel,
 } from '@components/Portfolio';
 import { LabelledOrders } from '@libs/types/OrderTypes';
+import { LabelledTracers } from '@libs/types/TracerTypes';
 
 interface OProps {
+    tracers: LabelledTracers;
     positions: Tracer[];
     holdings: Tracer[];
     allFilledOrders: LabelledOrders;
 }
-const Overview: FC<OProps> = ({ positions, holdings, allFilledOrders }: OProps) => {
+const Overview: FC<OProps> = ({ tracers, positions, holdings, allFilledOrders }: OProps) => {
     const { account } = useWeb3();
     const [currentPortfolio, setCurrentPortfolio] = useState(-1);
     const [currentPNL, setCurrentPNL] = useState(1);
@@ -82,7 +84,7 @@ const Overview: FC<OProps> = ({ positions, holdings, allFilledOrders }: OProps) 
                 {!account ? (
                     <ConnectOverlay />
                 ) : positions.length === 0 ? (
-                    <PositionOverlay tracers={positions} showMarketPreview />
+                    <PositionMarketOverlay tracers={tracers ?? {}} showMarketPreview />
                 ) : null}
             </HScrollContainer>
         </VScrollContainer>
